♻️ 팔로우 바텀시트 컴포넌트 한 파일로 병합 (#23)
